fix with resouce leak and switch to bigDecimal

+++ src/main/java/com/example/payment/PaymentApi.java
@@ -1,5 +1,7 @@
 package com.example.payment;
 
+import java.math.BigDecimal;
+
 public interface PaymentApi {
-    PaymentResponse charge(String apiKey, double amount);
+    PaymentResponse charge(String apiKey, BigDecimal amount);
 }
